# Make functional tests run faster