@pmeinhardt 2018-02-17
I have not found a satisfying answer to the second question, but I am okay with that for now (I you have any ideas, shoot). Plus I have an image of @tessi in the back of my head who is whispering – with a ghost like voice: "Smaller PRs, smaller PRs…" 👻 

So, here ya go 🚶