@pmeinhardt 2018-02-19
@klappradla, thanks a lot for the review :metal: 

As for the open question: My idea was protecting the developer machine by preventing modification of the local filesystem rather than isolating tests from other tests.

The tests now run in parallel because we don't mutate shared state (no files used by multiple test processes and no changing of the system process's current working directory), so that problem is solved for now :white_check_mark:  However, an error in a test case could still remove files on your local fs or create them in places where you don't want them… :bomb: 

I was therefore :thinking: whether we should run the (functional) test suite *as a whole* within some kind of isolated environment.

It would be ideal, if there was a mock for Elixir's `File` that performed *read* operations against the Disk (or an overwriting entry in the memory layer if there was a write to that location) but *write* to a memory-based filesystem… similar to what you can use OverlayFS, UnionFS and AUFS for. In the end though, this doesn't exist as far as I know.

The other solutions depend too much on what kind of system you use for development, except for Docker. With Docker though, we would have to run Docker containers within Docker, which works but means some more setup…

**Long story short**, I think you raised a really good point there: "close enough to reality"… To run the tests on different dev machines (OSes, filesystems…) definitely has its merits and maybe we don't even want to go down that other road.

Thank you for sharing your helpful thoughts!
(…and sorry for the lengthy brain dump)