@klappradla 2018-02-19
Oh ok, now I get your point 💡 

Well, I'm not familiar with any of the tools you list (except for Docker which I still don't think is a good option here).. A memory based file system could be a good option here though - however, it might come with an extra hassle for dev-setup which is again sort of a downer.

I personally always had good experience with just using temp-files (and directories) - did this last at Liefery if I remember correctly. The way I understood this was: as long as you fully stay in the `tmp` area, you won't really mess with anything and - especially if you have an *always-teardown™* you also won't pollute the local disc to hard 🚮